ocr-module/requirements.txt
pytesseract
Pillow
pdf2image
opencv-python
numpy